feat: Update intersection_with
Fixes an error in which `ZeroDivision` error is not triggered when it should be. Fixes it by testing for parallelism and sameness first. If the two lines are parallel, the function returns `None`. If they are the same, it returns `self`. If neither of those conditions are met, then it performs the calculations, making the `ZeroDivisionError` exception not necessary.